refactor(query): move MAX_SYNC_PAGE_SIZE into constants.ts
