+++ src/database/classes/QueryDb/library/constants.ts
