+++ src/database/classes/QueryDb/library/index.ts
@@ -1,4 +1,4 @@
 export * from './compileCondition';
 export * from './collectConditionColumns';
 export * from './assertIndexedColumns';
-export * from './MAX_SYNC_PAGE_SIZE';
+export * from './constants';
